NIJO-306 small refactoring
