+++ src/core/routes.js
@@ -20,7 +20,8 @@ import accountInvite              from "pages/account/accept-invite";
 import userBookingManagement      from "pages/user/booking-management";
 import accountStatement           from "pages/user/account-statement";
 import userInvite                 from "pages/user/create-invite";
-import personalInfo               from "pages/user/personalInfo";
+
+import personalInfo               from "pages/settings/admin-settings";
 
 import contactUsPage              from "pages/common/contact";
 import termsPage                  from "pages/common/terms";
@@ -44,7 +45,7 @@ export const routesWithHeaderAndFooter = [
     "/user/booking",
     "/user/payment-history",
     "/user/invite",
-    "/user/settings",
+    "/settings/admin",
 ];
 export const routesWithFooter = [
     ...routesWithHeaderAndFooter,
@@ -78,7 +79,8 @@ const Routes = () => (
         <Route path="/user/booking"               component={userBookingManagement} />
         <Route path="/user/payment-history"       component={accountStatement} />
         <Route path="/user/invite"                component={userInvite} />
-        <Route path="/user/settings"              component={personalInfo} />
+
+        <Route path="/settings/admin"             component={personalInfo} />
 
         <Route path="/contact"                    component={contactUsPage} />
         <Route path="/terms"                      component={termsPage} />
