+++ styles/block/personal-info.sass
@@ -29,31 +29,6 @@
     margin: 35px auto 80px
     color: #fff !important
     font-size: 13px
-  &__tabs
-    display: flex
-    justify-content: center
-    margin-bottom: 90px
-    border-bottom: 1px solid rgba(216, 218, 220, 0.5)
-    width: 100%
-    a
-      font-size: 13px
-      line-height: 16px
-      text-align: center
-      text-transform: uppercase
-      color: #797F89
-      margin-right: 10px
-      padding: 13px 25px
-      position: relative
-      &.active
-        &:after
-          content: ''
-          height: 4px
-          width: 100%
-          background: linear-gradient(90deg, #F9A51B 0%, #FECE08 100%), #C4C4C4
-          border-radius: 4px
-          position: absolute
-          bottom: 0
-          left: 0
   &__form
     .row
       margin-bottom: 18px
