+++ styles/blocks.sass
@@ -13,6 +13,7 @@
 @import "block/account"
 @import "block/error-page"
 @import "block/personal-info"
+@import "block/users-pages"
 
 .block.document
     h1
